fixed elements height in list by setting a size view changed collection_element root to LinearLayout because RelativeLayout fails to resolve the correct sizes added update_hook for template customization
fixed bug converting long from java
changed dynamicview id and timer id to be long (64bit)
generating dynamicview id randomly to avoid clashing across different runs
added helper attributes
added deleting attribute by setting it to None
added size unit conversion
reset package version

added 'px' unit
fixed GridView item size

+++ app/src/main/java/com/appy/Attributes.java
@@ -31,10 +31,10 @@ public class Attributes
     {
         static class Reference
         {
-            public int id;
+            public long id;
             public Type type;
             public double factor;
-            public Reference(int id, Type type, double factor)
+            public Reference(long id, Type type, double factor)
             {
                 this.id = id;
                 this.type = type;
@@ -80,7 +80,7 @@ public class Attributes
                 {
                     JSONObject referenceObj = referenceArray.getJSONObject(j);
 
-                    Reference ref = new Reference(referenceObj.getInt("id"), Type.valueOf(referenceObj.getString("type")), referenceObj.getDouble("factor"));
+                    Reference ref = new Reference(referenceObj.getLong("id"), Type.valueOf(referenceObj.getString("type")), referenceObj.getDouble("factor"));
                     references.add(ref);
                 }
